fixing array definitions

+++ finquant/monte_carlo.py
@@ -119,14 +119,14 @@ class MonteCarloOpt(MonteCarlo):
              list of [expected return, volatility, sharpe ratio].
         """
         # select random weights for portfolio
-        weights = np.array(np.random.random(self.num_stocks))
+        weights = np.array(np.random.random(self.num_stocks), dtype=float)
         # rebalance weights
         weights = weights / np.sum(weights)
         # compute portfolio return and volatility
         portfolio_values = annualised_portfolio_quantities(
             weights, self.return_means, self.cov_matrix, self.risk_free_rate, self.freq
         )
-        return (weights, np.array(portfolio_values))
+        return (weights, np.array(portfolio_values, dtype=float))
 
     def _random_portfolios(self) -> Tuple[pd.DataFrame, pd.DataFrame]:
         """Performs a Monte Carlo run and gets a list of random portfolios
